advise/seccomp: Move non k8s related functions to tracer pkg
